+++ src/main/java/core/bassyntax/model/Cat.java
@@ -1,0 +1,13 @@
+package core.bassyntax.model;
+
+public class Cat {
+    private int age;
+    private String name;
+    private String ownerName;
+
+    @Override
+    public String toString() {
+        return "Cat{" + "age=" + age + ", name='" + name + '\''
+                + ", ownerName='" + ownerName + '\'' + '}';
+    }
+}
