+++ src/main/java/core/bassyntax/model/Dog.java
@@ -1,0 +1,4 @@
+package core.bassyntax.model;
+
+public class Dog {
+}
