+++ src/test/java/core/bassyntax/HelloWorldTest.java
@@ -1,4 +1,4 @@
-package core.basesyntax;
+package core.bassyntax;
 
 /**
  * Feel free to remove this class and create your own.
